requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==1.2.0
Django==4.0.4
django-dotenv==1.4.2
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.2
whitenoise==6.2.0
